Changed: move the AuthService to ./core
