+++ src/app/components/account-activation/account-activation.component.ts
@@ -1,5 +1,5 @@
 import {Component, OnInit, OnDestroy} from '@angular/core';
-import {AuthService} from "../../services/auth.service";
+import {AuthService} from "../../core/auth.service";
 import {ActivatedRoute} from "@angular/router";
 import {Subscription} from "rxjs/Rx";
