+++ src/app/components/experience-block/experience-block.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import {User} from '../../classes/user';
-import {AuthService} from '../../services/auth.service';
+import {AuthService} from '../../core/auth.service';
 
 @Component({
   selector: 'xp-experience-block',
