+++ src/app/components/header/header.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { AuthService } from "../../services/auth.service";
+import { AuthService } from "../../core/auth.service";
 
 @Component({
   selector: 'xp-header',
