+++ src/app/register/register.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { Subscription } from 'rxjs/Rx';
 
 import { User } from '../classes/user';
-import { AuthService } from '../services/auth.service';
+import { AuthService } from '../core/auth.service';
 
 @Component({
     selector: 'xp-register',
